# Faster X-Rotations in C