@rsln-s 2024-08-20
LGTM! Here's some benchmarking data:

```
jpmorganchase:performance_c
For N = 12, LABS finished in 0.0028 on average, min: 0.0026, max: 0.0035
For N = 16, LABS finished in 0.1354 on average, min: 0.1228, max: 0.1461
For N = 20, LABS finished in 0.1559 on average, min: 0.1441, max: 0.1769
For N = 24, LABS finished in 0.9767 on average, min: 0.9469, max: 1.0068
For N = 28, LABS finished in 19.5529 on average, min: 19.4360, max: 19.6836
For N = 32, LABS finished in 355.0791 on average, min: 350.2837, max: 361.5040
For N = 33, LABS finished in 717.4631 on average, min: 710.2043, max: 726.1118
------
mebenstein-vorticity:performance_c
For N = 12, LABS finished in 0.0013 on average, min: 0.0012, max: 0.0016
For N = 16, LABS finished in 0.1293 on average, min: 0.1180, max: 0.1483
For N = 20, LABS finished in 0.1510 on average, min: 0.1325, max: 0.1745
For N = 24, LABS finished in 0.7461 on average, min: 0.7184, max: 0.7760
For N = 28, LABS finished in 11.4725 on average, min: 11.3852, max: 11.6067
For N = 32, LABS finished in 234.7204 on average, min: 234.3182, max: 235.0045
For N = 33, LABS finished in 482.1637 on average, min: 476.6924, max: 485.6158
For N = 34, LABS finished in 958.0517 on average, min: 946.1425, max: 968.6004
```

Overall, looks like a nice speedup